Pscheck reasks until it is safe to install
The logic is now entirely in pscheck, and moved to the individual functions
instead of confirm (because create, which calls confirm, does not need this).
It ask to proceed, but if there are still processes open, it will check again
and ask again.  You can force an install by passing --yes, which will still
print the warning, but not stop the installation.