conda skeleton pypi: Automatically get dependencies and entry points
This only works if they are specified using distribute (i.e., install_requires
or entry_points in setup() in setup.py). To do this, it downloads the source
and runs setup.py in an environment that patches distutils.core.setup to
export the dependencies and entry points.

I also automatically added tests for the entry points if there are any
(entry-point --help), to help promote good practices.