First attempt at a fix for conda install in Python 3
This hopefully fixes #268 and #270, but I'm not too sure about the
solution. It's the only way I found that worked. It doesn't seem possible to
encode arbitrary strings using .encode(), because no encoding supports every
possible byte. So if you want to deal with data as bytes, you have to open it
as bytes.

This needs to be thoroughly tested, both in Python 2 and Python 3.