Allow the solver to tell the plan module what packages to remove
This is triggered by adding 'remove ' to the front of the package name
returned by solve. This can be switched off with a flag (which will later be
False by default).

Also, clean up the code at the end of solve2 that pulls together the
solution. The long list comprehension was getting hard to read.