Workaround launch issue in Windows XP/Python 3
Under Python 3.4 and Windows XP (32 or 64 bit), the spawned process
does not inherit stdout/stderr. If the process then tries to spawn
its own subprocess, this raises an exception when Python can't
make a new file handle. This commit forces subprocess to spawn the
new process with inherited file handles, avoiding the issue.