Correctly avoid non-specified features, even if they are directly in the specs
Previously it would only prevent installing packages with features that
weren't explicitly asked for from packages listed in the specs, but it really
should do this for every package found, including dependencies. This came up
in Windows 32-bit, where there was a newer build of NumPy 1.6 with the mkl
feature than without, causing it to be chosen when installing a package that
depended on it.