Allow spaces to be contained inside conda run
This fixes and issue where this will not run:

    conda run --prefix=/path/to/env "python --version"

The issue is around the way arguments are passed to subprocess.Popen.
Popen can't handle `["python --version"]` as an argument, which is what
we're passing in, but it can handle `["python", "--version"]`.  Using
`shlex.split` gives us a shell-safe way of splitting a string so we
can turn `python --version` from the command line into the second list.