conda.install.symlink_conda: Check for symlink targets
Conda currently assumes a system-wide install:

  $ pip install --user conda
  $ command -v conda
  /home/me/.local/bin/conda
  $ conda create -n foo --yes python
  $ ls -l ~/envs/foo/bin/conda
  lrwxrwxrwx 1 ... /home/me/envs/foo/bin/conda -> /usr/bin/conda

I'm not sure if it's better to adjust:

  actions[SYMLINK_CONDA] = [config.root_dir]

or:

  root_dir = abspath(expanduser(os.getenv('CONDA_ROOT',
                                          rc.get('root_dir', sys.prefix))))

but regardless of how we get things working, we'll still want this
check to make sure we're not broken.

The exception isn't caught anywhere, so if it fails you get the whole
traceback and a half-setup ~/envs/foo/.  I think there should be some
"abort this command, remove the broken environment, and hide the
traceback unless debugging is enabled" handling that catches it, but I
don't know the code base well enough to know where that would fit (it
may even already exist).