conda.install.symlink_conda: Link to the current executables
Before this commit we had:

  actions[SYMLINK_CONDA] = [config.root_dir]

where:

  root_dir = abspath(expanduser(os.getenv('CONDA_ROOT',
                                          rc.get('root_dir', sys.prefix))))

and we assumed the conda executables (conda, activate, and deactivate)
were in CONDA_ROOT/bin/.  In fact, the install should succeed if the
executables are anywhere in the caller's PATH.  For example, they may
be in ~/.local/bin because the user installed conda with:

  $ pip install --user conda

This commit drops the root_dir argument from symlink_conda and uses
shutil.which [1] to figure out where to point the symlink (falling
back to distutils.spawn.find_executable [2,3] for Python <3.3).

There are also some tweaks in conda.plan to support commands that
don't need arguments (which is what SYMLINK_CONDA is now).

We use os.realpath() to dereference symlinks in the root paths so
something like:

  $ conda create -n foo --yes python
  $ source activate foo
  $ conda create -n bar --yes python

will create ~/envs/bar/bin/conda (and similar) linking to the original
~/.local/bin/conda (or wherever) instead of having bar's conda link to
foo's link to the original.

[1]: https://docs.python.org/3.3/library/shutil.html#shutil.which
[2]: https://docs.python.org/3.3/distutils/apiref.html#module-distutils.spawn
[3]: https://docs.python.org/release/2.4/dist/module-distutils.spawn.html