Move all of the error messaging over to activate/deactivate
All of the sys.stderr writing that happening inside the various commands
was only being used in the POSIX world as the .bat files do all of their
own error reporting.

Since these commands follow as sub-commands, the parser was grabbing
ahold of various commands and displaying its own error messages instead
of the custom ones.  This commit moves those messages out to the shell
side of things and uses exit codes to communicate back and forth.

One note, I saw some oddness in the PATH where `$HOME/anaconda/bin` was
ending up inside the PATH variable.  Since the most important thing was
to check the first value, I've written code that does just that and
ignores anything else that makes it into the path by virtue of the shell
starting on different local environments.