Add support for activate/deactivate scripts
This provides a mechanism for packages to run scripts on activate and
deactivate.  The idea is that a script would provide a startup and
shutdown script that starts something when activated and stops it when
deactivating.  The use case I've been working from is supervisord, which
would also start other packages that are installed *and* provide a
mechanism for inspecting the state of those running applications.