Refactor how root_dir is determined
This moves from inspecting the current running Python to using
convention to figure out where the root directory is located.  This
works as-is with existing Anaconda / Miniconda installations and
provides a mechanism for falling back to sys.prefix.

Eventually, we should check for that `sys.prefix` is in the default
locations, and if not, prompt users to set their configuration to point
to their non-standard location.