Refactor to use new conda.packages.from_file()
This addresses the issue I ran into with conda install --offline and the
offline index requiring depends.  I've verified that with this change
you can install the packages that were in [issue 236][1] by manually
editing the zlib meta data, re-tarring the file, then attempting to
install it into a clean environment.

[1]: https://github.com/ContinuumIO/anaconda-issues/issues/236