Fix issue related to activate environments without Python (v1.4.1)
For some reason this code only checked to see if there was a
`python.exe` file in the environment rather than checking to see if
something that's always there is present.  This means that any
environment that does not contain one of the Python packages will not
allow activation.

This change also brings back building of OS X packages of conda-env as
the new OS X build box is back online.