Fix weird issue where $0 == "-/bin/bash"
I've got an issue, probably related to my highly customized bash prompt,
where $0 == "-/bin/bash".  That cases errors when passed to `basename`
because thinks that's a parameter and tries to treat it as such.  This
modifies the code to ignore those errors.  The path that its attempting
to detect is a direct call to "activate" or "deactivate" which won't
trigger those errors.