Swaps out conda-meta/{{ pkg }}.json to conda-meta/{{ pkg }}/index.json
This is a pretty big change and this is simply the proof of concept to
have a conversation around the code.  This handles environment
creation/installation and removing a package.

Some investigation needs to be done into various interaction with the
`shutil` code on Windows.  There are known edge-cases that are handled
in other parts of the code regarding `shutil.rmtree`.  We need to make
sure that those are handled properly.

Also note that all tests still pass.  There are no tests explicitly
verifying that the correct `{{ pkg }}.json` file is created, but I'm
unsure whether there are tests that verify that things like `conda list`
continue to work.  We should definitely add tests that verify the
expected behavior that work on master and use those to verify that the
commands continue to work here.

Still to do:

* Track how the package was brought into the environment
* Track the origin channel for a given package and ensure the
  environment has that meta data.