Add binstar's conda channel to the defaults
This means that anything that's uploaded to binstar's main conda channel
is available by default to all packages.  This simplifies the release of
conda, conda-env, and conda-build as anyone with binstar access can make
a release.

Note: this requires that a package be in the `main` channel of the conda
organization in order to get picked up.