include cwd when searching for additional commands
This patch includes the current working directory when searching for external
commands.

For example, assume there is a file in the current working directory, named
'conda-hello' that simply echos 'Hello Conda!'. Before this patch one needed:

```
  $ PATH=. conda hello
  Hello Conda!
```

Now one can simply do:

```
  $ conda hello
  Hello Conda!
```

The current working directory is obtained in a platform independet way and is
placed in front of the entries in PATH to give commands in the current working
directory precedence. This behaviour mimicks common Python, where the current
working directory is always searched for modules. However, it differs from git
(from which this extension mechanism was presumably inspired)
which doesn't include the current working directory.