Resolve module falls back to min_sat(alg='sorter') if not all solutions were found
Previously it would iterate the first 1000 solutions, but if there were many
more, it might not find the minimal one. Raising the maximum iteration count
makes it very slow (an example I tested required ~70000 iterations). Instead,
we fallback to the sorter algorithm if there are more than 100 solutions,
which appears to be faster than BDD (this agrees with the heuristic described
in the code in 1c7f84ca065c53ff8cd20a0370eb2a6a06dcb1a4, since the equation
has all coefficients of 1).