Fixes issue on Windows where forced update fails partial update
Talked with @jreback at pycon about this.  Users are leaving IPython
Notebook sessions open with `pandas` imported, then trying to update
pandas.  Attempting a force install when prompted to continue causes the
installation to partially update.

The partial update is caused when conda attempts to relink files that
are open.  Some of them work, other's don't.  Unfortunately, we don't
know that until we attempt to write to them.

The try/except a few lines down that catches OSError is swallowing the
`WindowsError(5, "Access is denied")` that is being thrown by
`os.unlink`.  Simply modifying that, however, causes an issue where some
of the files would be removed before we know that we can't update.

Fixes issue #1239
Related to https://github.com/ContinuumIO/anaconda-issues/issues/256