Use minimal_unsatisfiable_subset() in guess_bad_solve()
This is much faster than the previous algorithm. We can now use it with any
size of specs, since it is a bisection that takes roughly log2(n) steps.

The downside is that we no longer get all minimal bad combinations, but rather
only one (which may not even be the smallest; a "minimal" subset here means it
cannot be shrunk, not that it is smaller than every other bad subset).
However, since every bad subset typically has to be addressed eventually,
addressing one at a time is not that bad, especially since the hints are now
so fast to generate.

Also give more helpful messages with the hints.