Preserve old links with new navigation
As a messy compromise, we're going to leave source/build.rst where it
is instead of moving it to source/building/recipe.rst as we really should,
so that links from outside the repository will still work. At the same
time, we make the navigation look right and behave as it should.