Don't fetch the package metadata from the network with conda remove
The index cache is now used by default. The index is needed only to display
the channel information when show_channel_urls is set to True. The worst that
will happen when there is no cache present is that the channel will show as
"<unknown>" instead of the actual channel. This makes conda remove much
faster.

The index can still be forcibly fetched with conda remove
--no-use-index-cache.

Fixes #1667.