Always attempt to delete files in `rm_rf`
Ask for forgiveness, rather than permission.

When creating a user environment based on a root environment, extra package files would be left in the root of the new user environment. `os.access` checks if the inode itself is writable, not if the directory containing it is writable (which is all that is needed to unlink a file).

Fixes conda/conda#1861