Manage BeeGFS hard-links non-POSIX configuration
This is a fix for #2345

BeeGFS do not support hard-link between files if different directories.
Depending on how BeeGFS is configured, a symbolic-link may be created instead
of a hard-link.

This commit modifies try_hard_link function to check the link created is really a
hard-link, not a symbolic link.