install: Make move_to_trash less frightening (#2587)
The relpath() calculation was only added as a debugging aid in
56539d, but it made a dangerous assumption that the path being
trashed was a subfolder of the root environment.

Now, people have been using it with folders that are not in the
same hierarchy at all and this leads to escaping the .trash dir
(and potentially the root prefix) so here we guard against this
security issue.

To do so, we normalize and remove any leading '..' from the path.
This has a nice side effect of preventing *any* '..' from making
their way to the final UNC trash path, which is required because
UNC paths must be absolute and normalized for Windows to process
them.

.. which fixes:
  https://github.com/conda/conda/issues/2582

.. and gets to the bottom of the code smell in:
  https://github.com/conda/conda/issues/2579