updated activate.bat and deactivate.bat
Addressed issues with the Windows code versions not conforming to the
style of the recently updated Unix codes.

Consequently these changes revealed certain weaknesses in the unix code
that prompted slight updates that allows the activate/deactivate
scripts to fully harness the existing design logic of the larger Conda
platform. As an example, the CONDA_VERBOSE environment variable is
borrowed by the scripts and depending upon whether or not they were
simply set or exported prior to running the activate/deactivate scripts
will have different results.

Added a README file to explain in depth the design logic of the unix
and windows scripts.

Tests have yet to be updated/checked in relation to the Windows .bat
codes.