addressed conda-build test failure, added no-brainer ksh support
Identified condo-build test failure. Was a combination of the
deactivate script exiting with a non-zero status if there was nothing
to deactivate and also the activate/deactivate script’s reliance on the
CONDA_DEFAULT_ENV variable to detect whether or not an environment was
active (this didn’t work because condo-build would set this for it own
internal use). These issues caused the activate to improperly activate
the build environment.

As a result of how condo-build uses activate/deactivate, added unittest
that ensure that both activate/deactivate runs without non-zero exit
statuses in success cases (previously various parts would return
non-zero as part of a cross shell support for a makeshift if logic).

Also added KSH support. This was a complete no-brainer as the existing
bourne shell scripts were entirely valid for KSH, just needed to add
the necessary TravisCI testing and the clauses in whichshell* scrips.