Adding PowerShell support and Windows Cygwin/MinGW/MSYS2 support
Added preliminary PowerShell support. This is not tested using
unittesting. Primarily since PowerShell’s style of controlling the
prompt is significantly more complex and will need work that is outside
of the scope of the goals of this branch.

Tested against Cygwin, MinGW, and MSYS2. Added minor tweaks to
whichshell_ps.bash to address the variations in their respective `ps`
commands compared to Unix. The majority of AppVeyor tests will pass.