Created an environment trimming function
Initial commit of environment trimming function that processes the
activating environment to decide whether it is located in the env_dirs
in which case only the base name is used otherwise the environment is
resolved into a full path.

This is used to simplify the prompt such that when an environment is
activated by its name it will look the same as when it is activated by
its path.