version: Fix comparision of multi-part bits
if VersionOrder('4.3.2.dev2+38bb992b') < VersionOrder('4.3.2'):
    raise RuntimeError("Non-native subdir support only in conda >= 4.3.2")

The problem here was that each part of [0,'DEV',2] was being compared
against 0 and strings ('DEV') are considered less than ints (0).

Also re-implement __eq__ in terms of __lt__ only to prevent repeated logic
and remove the self.fillvalue thing (we always use 0 or 'DEV' now). Though
if anyone cares it could be supported.