{de,}activate: Limit scope of `IFS` to `find` command only
`IFS` was set so that the output of the `find` command that looks
for package {de,}activation scripts could be used in a for loop
(since they must be `source`ed we cannot use `find -exec` to run
them).

Unfortunately this change would remain active afterwards and
this meant that normal variable delimiting did not work, so if a
script set CFLAGS="flag1 flag2", that value would be treated as
a single argument when passed to the compiler.