{de,}activate: Remove BSD-ism from sed calls
The `-E` flag only works on macOS (and some *BSD I guess). In GNU sed, the flag is
`-r`. These flags turn on 'extended regular expressions'.

Extended regular expressions were only needed in one of the three invocations (the
other case where initially it looked like it serves a purpose removed the `:` only
to put it back again.

The case that did make use of extended regular expressions is rewritten to use sed
'chaining' instead. I could've added code to determine which sed is being used but
macOS sed does not provide a `--version` flag and even if it did, calling it twice
would add the overhead of another process invocation.