link: Pass target_prefix as env_prefix regardless of is_unlink
Existing pre-link scripts expect $PREFIX to be the destination
prefix into which package is being installed.

In 4.3 that code changed, None gets passed as env_prefix so
prefix gets used instead which is now the package cache and
not the destination. The commit which made this change is:
95241d2c6e19fd8a3223e3c703a39db8a9dcdb8f

This bug can be reproduced by:

```
conda install conda=4.3
conda install -c continuumcrew anaconda-navigator
```

resulting in:

```
/home/ray/mc-x64-3.5/pkgs/anaconda-navigator-1.4rc2-py_1/bin/.anaconda-navigator-pre-link.sh: line 3: /home/ray/mc-x64-3.5/pkgs/anaconda-navigator-1.4rc2-py_1/bin/python: No such file or directory
```