Detect and remove stale lock files (#4862)
When a lock file is found, attempt to detect if the process that
created it is still running. If not, consider the lock file stale
and remove it. Be somewhat pessimistic about it in the sense that
if the result is inconclusive, assume the lock file is still active.

This works by adding the name of the host that created the file into
the filename of the lock file in addition to the PID of the process
that created it. When a lock file is encountered and the host name in
the filename matches the current host name, and a process with the
PID found in the lock filename is not running on the local system,
then declare the lock stale and remove it.

This makes conda clean up stale lock files automatically in the
majority of cases, while in other cases (reused PID numbers,
shared file system), one can still run `conda lock --clean` manually.