Remove matching to an existing package cache for explicit install
When writing the `urls.txt` entry for an explicit install, do not attempt
to match it to an existing package cache. If we do this, the package cannot
be installed explicitly anymore because the `Dist` object that is recorded
in the `urls.txt` is different from the one that the explicit install
attempts to install (channel is different).