Fix exception when generating JSON output
When running the command `conda create -n <env> --dry --json python`, the
command would fail with `AttributeError("'PrefixSetup' object has no
attribute 'link_dists')`.

This PR fixes this by using the `link_precs` and `unlink_precs` attributes
instead. It also removes the `unlink_link_transaction` key from `kwargs`
because otherwise the JSON serializer would complain that "an object of
type `UnlinkLinkTransaction` is not JSON serializable."

It's not fully clear to me what the desired JSON output should be. In
conda 4.2.x, it used to be

```
"actions": {
  "LINK": [
    "defaults::vs2015_runtime-14.00.23026.0-0 1",
    ...
  ]
}
```

In conda 4.3.x, it used to be

```
"actions": [
  {
    "LINK": [
      {
        "base_url": null,
        "build_number": 0,
        "build_string": "0",
        "channel": "defaults",
        "dist_name": "vs2008_runtime-9.00.30729.5054-0",
        "name": "vs2008_runtime",
        "platform": null,
        "version": "9.00.30729.5054",
        "with_features_depends": null
      },
      ...
    ]
  }
]
```

Now in conda `4.4.x`, the output is back to what it was in conda `4.2.x`.
From the code that generates the JSON output it looks like this is intended
and not a byproduct of this PR.