Fix msys2 activate and deactivate
Mostly deactivate was returning a Windows PATH to the
MSYS2 shell, but also, the conversion was hand-rolled
and incorrect; it did not respect /etc/fstab. Instead
use cygpath.exe (or CYGPATH env var).

Also, the conversions were being done incorrectly. We
need to convert to or from the shell depending on the
context.