Drop custom copy + remove fallback for rename failures
os.rename() returns EXDEV for directories and EINVAL for files in a docker
container with overlayfs or aufs storage backend.

xref: https://github.com/moby/moby/issues/25409#issuecomment-238537855

The fallback for this was to use copy + remove, but it assumed that the rename
failed for a file. So, whenever a copy attempt was made on a directory, it
cascaded into an uncaught EISDIR.

This patch delegates the copy+remove logic to shutil.move() which works for
directories and files and saves us from reinventing the wheel.