Fix for envvars being carried over to subshell
The tests appeared to be running into an issue where when subshelling from bash to csh (for example). The conda environment variables were being carried over from bash to csh and hence messing with what conda activate thought it needed to do. Without rethinking everything this is a reasonable fix that requires us to rely on both CONDA_PREFIX and CONDA_SHLVL to decide whether we are activating or reactivating.