Fix MSYS2 path conversion in the face of bind mounts and /etc/fstab
There is a subtle issue that is caused by Cygwin/MSYS2's fstab where it
contains a bind mount /bin => /usr/bin. With cygpath in /Library/usr/bin
(translated to /usr/bin) this bind mount is translated to /bin and thus
/Library/bin (contains native Windows software) is shadowed by
Library/usr/bin (contains MSYS2 software). To address this we translate
the prefix root and then just append the individual paths to the result
of that thus preventing this unfortunate fstab-related short-circuiting.