clarify conditions for timestamp minimization
Timestamp minimization is only necessary when the SAT problem has multiple
solutions, that is to say when it has not converged to a single solution.
Replace the has_multiple_simple_solutions function with the is_converged
function.  This function has a docstring which explains the condition being
tested for and how this test is performed.