reduce broadening of spec scope during get_reduced_index
get_reduced_index does a recursive collection of entries in repodata.  This
collection can be much smaller if we start by considering as many exact specs
as possible, and limit conda from broadening the search when a more exact match
is already in place.  This commit does so by first sorting specs by the number
of exact specs they in their own dependencies, then by taking that into account
with a pretty simple check for whether build string or version is already
present in the expanded collection, and omitting specs that would broaden the
search space.