Wrap more calls to conda because it may not be on PATH properly
And make all MS-DOS commands uppercase BECAUSE THIS WAS THE HEIGHT
OF TECHNOLOGICAL SOPHISTICATION BACK IN THE OLDEN DAYS AND DID NOT
MEAN SHOUTING LIKE IT DOES NOW.

Ensure the first entry added to PATH doesn't end with a slash.

Carefully remove these added-to-front-of-PATH entries from any
env. vars set in batch files that conda emits.