Quote path to temporary requirements file
If there is a space in the path to the .yml file for an environment, conda create will fail. This is because the temporary requirements file created here will be created in the same directory as the .yml file, and when the path to it is given to pip, it will fail if there is a space in it and the full path isn't quoted. This commit is ensuring that the path is quoted as part of the pip command.