Rework how capturing stdio and stderr is done
The way it was, it was not possible to print Unicode (cStringIO), fixing that
led to many knock-on issues. Now we use StringIO and override the write function
so we can handle conversions ourselves (and if we decide we want to enforce
a single string type (unicode on Py2 and str on Py3) later this will help us
to determine (via breakpoints) where other types (bytes) are slipping in.