Various bits around testing
1. `--basetemp` option.

The most important/useful is the use of pytest's tmpdir stuff. This
means that you can pass, for example:

`--basetemp=${HOME}/conda-tmp`

.. I added this because by default, TMPDIR gets used and that's a
different device from where I run conda. This meant that my tests
all fell back to copying, making them slower (and less like most
of our users setups, though really we need to use pytest fixtures
so we run some of our tests with `hardlink` and `copy` settings.

As well as this, you can set `CONDA_TEST_TMPDIR` as an env. var,
but really, why would you. I may remove this.

2. `CONDA_TEST_SAVE_TEMPS` env. var.

Should probably be a fixture too.
.. Set this to prevent deletion of envs created by `make_temp_env`
and scripts made by `wrap_subprocess_call`.

3. Improve the `run_command` / `wrap_subprocess_call` argument
processing a little. This allows for more natural expression of
tests in shell, batch and python, at the cost of some 'custom'
behaviour we should document:

```
When a single multiline argument is passed it is copied verbatim.
Multiline arguements that are mixed with more arguments are not
passed verbatim but are quoted according to some specific rules.
`'` and `"` are used for quoting as deemed appropriate.
```

It may end up that we'll want a way of specifying what to replace
newlines with should be for each language? invocation of
`run_command`? For example with Python we *could* replace newlines
with `;` and then quote that. Luckily Python seems happy with
multiline arguments though (at least so far, on macOS!)