Add a --dev mode to wrap_subprocess_call
This was added for test_init_dev_and_NoBaseEnvironmentError but it should be generally
useful. It changes the temporary invocation scripts to set `conda_exe` to
`['prefix/bin/python', '-m' 'conda']` so that run_script() can launch conda development
sources (by virtue of PWD being on sys.path) from a Python in located elsewhere (usually
a different prefix).

To make it clear, the code is:

```
        if dev_mode:
            conda_exe = [abspath(join(root_prefix, 'bin', 'python')), '-m', 'conda']
        else:
            conda_exe = [environ.get("CONDA_EXE", abspath(join(root_prefix, 'bin', 'conda')))]
```

.. so it is important to realize that when dev mode is used, the `CONDA_EXE` env. var. is
ignored. I may be prudent to add a warning when they get out of sync so we can fix that.