Fix {,t}csh test_activation
It only worked if run from a specific folder. conda.csh was broken, basing things off of
CONDA_ROOT which is deprecated now. Instead, in this fallback case (where conda is not a
shell alias) we instead determine the prefix from _CONDA_EXE alone.

Also change the shell.args stuff so that the args from the baseshell are combined with
those from the shell itself, with the baseshell ones appearing first - many programs allow
flags seen later to override those seen earlier so this fits with that - not that we use
this code at all (I tried to add some `-x` stuff while debugging this but our shell comms
stuff fell over (pexpect).